Rename storage-limits endpoint to limits
